Improve sp-mantis2github script: version 1.3
Be more liberal in detecting migrated issues: the "Originally reported as"
GitHub Issue comment need not be the first comment.  This allows several Mantis
issues to be manually marked as migrated to a single GitHub issue by adding a
properly formed note to the Mantis issue and properly formed comment to the
GitHub issue, and relinking will respect the markings.